analyze: rewrite: move build_span_index into unlower module
